ingester/requirements.txt
redis==5.0.1
faker==22.5.1
python-dotenv==1.0.0
asyncio